Fixed: Backend Updates from Sonarr

+++ src/NzbDrone.Core/Tags/TagService.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using NzbDrone.Core.Datastore;
 using NzbDrone.Core.Messaging.Events;
 using NzbDrone.Core.Notifications;
 using NzbDrone.Core.Profiles.Delay;
@@ -120,7 +122,12 @@ namespace NzbDrone.Core.Tags
 
         public Tag Add(Tag tag)
         {
-            //TODO: check for duplicate tag by label and return that tag instead?
+            var existingTag = _repo.FindByLabel(tag.Label);
+
+            if (existingTag != null)
+            {
+                return existingTag;
+            }
 
             tag.Label = tag.Label.ToLowerInvariant();
 
@@ -142,6 +149,12 @@ namespace NzbDrone.Core.Tags
 
         public void Delete(int tagId)
         {
+            var details = Details(tagId);
+            if (details.InUse)
+            {
+                throw new ModelConflictException(typeof(Tag), tagId, $"'{details.Label}' cannot be deleted since it's still in use");
+            }
+
             _repo.Delete(tagId);
             _eventAggregator.PublishEvent(new TagsUpdatedEvent());
         }
